+++ examples/aggregate_merging/src/projectors.rs
@@ -1,13 +1,10 @@
-use std::sync::Arc;
-
 use async_trait::async_trait;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 use sqlx::{Executor, Postgres, Transaction};
-use tokio::sync::Mutex;
 use uuid::Uuid;
 
-use esrs::projector::PgProjector;
+use esrs::projector::Projector;
 use esrs::store::StoreEvent;
 
 use crate::structs::{CounterError, ProjectorEvent};
@@ -25,7 +22,7 @@ pub struct CounterProjector;
 // model - in this case, using queries which only update count_a when EventA is received, and count_b
 // when EventB is received, would be sufficient to guarantee soundness.
 #[async_trait]
-impl<T: Clone + Into<ProjectorEvent> + Send + Sync + Serialize + DeserializeOwned> PgProjector<T, CounterError>
+impl<T: Clone + Into<ProjectorEvent> + Send + Sync + Serialize + DeserializeOwned> Projector<Postgres, T, CounterError>
     for CounterProjector
 {
     async fn project(
@@ -123,43 +120,5 @@ impl Counter {
             .fetch_optional(executor)
             .await
             .map(|_| ())
-    }
-}
-
-// Here's an example of how one might implement a projector that is shared between two aggregates, in a
-// way that guarantees soundness when both aggregate event types can lead to modifying a shared
-// column in the projection. Note that this requires a different Aggregate setup, where a shared
-// (inner) projector is constructed, two SharedProjectors are constructed to wrap the inner, and then
-// an EventStore for each event type is constructed and passed the right SharedProjector for it's Event type,
-// before being passed into some Aggregate::new. It also, requires cloning every event
-pub struct SharedProjector<InnerEvent, InnerError> {
-    inner: Arc<Mutex<dyn PgProjector<InnerEvent, InnerError> + Send + Sync>>,
-}
-
-impl<InnerEvent, InnerError> SharedProjector<InnerEvent, InnerError> {
-    pub fn new(inner: Arc<Mutex<dyn PgProjector<InnerEvent, InnerError> + Send + Sync>>) -> Self {
-        SharedProjector { inner }
-    }
-}
-
-#[async_trait]
-impl<Event, Error, InnerEvent, InnerError> PgProjector<Event, Error> for SharedProjector<InnerEvent, InnerError>
-where
-    Event: Into<InnerEvent> + Clone + Send + Sync + Serialize + DeserializeOwned,
-    InnerEvent: Send + Sync + Serialize + DeserializeOwned,
-    InnerError: Into<Error>,
-{
-    async fn project(
-        &self,
-        event: &StoreEvent<Event>,
-        transaction: &mut Transaction<'_, Postgres>,
-    ) -> Result<(), Error> {
-        let event: StoreEvent<InnerEvent> = event.clone().map(Event::into);
-        let inner = self.inner.lock().await;
-        let result = inner.project(&event, transaction).await;
-        match result {
-            Ok(()) => Ok(()),
-            Err(e) => Err(e.into()),
-        }
     }
 }